Move some kernel headers
