+++ include/psp2kern/jpegenc.h
