+++ include/vitasdkkern.h
@@ -9,29 +9,27 @@
 #include <psp2kern/fios2.h>
 #include <psp2kern/gpu_es4.h>
 #include <psp2kern/idstorage.h>
+#include <psp2kern/jpegenc.h>
+#include <psp2kern/netps.h>
 #include <psp2kern/npdrm.h>
 #include <psp2kern/pfsmgr.h>
 #include <psp2kern/post_ss_mgr.h>
 #include <psp2kern/power.h>
 #include <psp2kern/registrymgr.h>
-#include <psp2kern/sblacmgr.h>
-#include <psp2kern/sblaimgr.h>
-#include <psp2kern/sblauthmgr.h>
-#include <psp2kern/syscon.h>
 #include <psp2kern/touch.h>
-#include <psp2kern/uart.h>
 #include <psp2kern/udcd.h>
 #include <psp2kern/usbd.h>
 #include <psp2kern/usbserial.h>
 #include <psp2kern/usbserv.h>
-
-#include <psp2kern/avcodec/jpegenc.h>
 
 #include <psp2kern/io/devctl.h>
 #include <psp2kern/io/dirent.h>
 #include <psp2kern/io/fcntl.h>
 #include <psp2kern/io/stat.h>
 
+#include <psp2kern/kernel/acmgr.h>
+#include <psp2kern/kernel/aimgr.h>
+#include <psp2kern/kernel/authmgr.h>
 #include <psp2kern/kernel/cpu.h>
 #include <psp2kern/kernel/debug.h>
 #include <psp2kern/kernel/debugled.h>
@@ -50,10 +48,12 @@
 #include <psp2kern/kernel/ssmgr.h>
 #include <psp2kern/kernel/suspend.h>
 #include <psp2kern/kernel/sysclib.h>
+#include <psp2kern/kernel/syscon.h>
 #include <psp2kern/kernel/sysmem.h>
 #include <psp2kern/kernel/sysroot.h>
 #include <psp2kern/kernel/systimer.h>
 #include <psp2kern/kernel/threadmgr.h>
+#include <psp2kern/kernel/uart.h>
 #include <psp2kern/kernel/utils.h>
 #include <psp2kern/kernel/kbl/kbl.h>
 #include <psp2kern/kernel/sysmem/data_transfers.h>
@@ -68,7 +68,5 @@
 #include <psp2kern/lowio/i2c.h>
 #include <psp2kern/lowio/iftu.h>
 #include <psp2kern/lowio/pervasive.h>
-
-#include <psp2kern/net/net.h>
 
 #endif /* _VITASDKKERN_H_ */
